Module rawr::options

Configuration options for API requests.

Structs

LinkPost
Options used when creating a link post. See structures::subreddit for examples of usage.

ListingOptions
Configures a paginated listing.

SelfPost
Options used when creating a self post. See structures::subreddit for examples of usage.

Enums

ListingAnchor
Used to 'anchor' the pagination so you can get all posts before/after a post.

TimeFilter
Used for filtering by time in the top and controversial queues.