Struct rawr::responses::comment::Comment

pub struct Comment {
    pub subreddit_id: String,
    pub banned_by: Option<String>,
    pub removal_reason: Option<String>,
    pub link_id: String,
    pub likes: Option<bool>,
    pub replies: Value,
    pub saved: bool,
    pub id: String,
    pub gilded: u64,
    pub archived: bool,
    pub author: String,
    pub score: i64,
    pub approved_by: Option<String>,
    pub body: String,
    pub edited: Value,
    pub author_flair_css_class: Option<String>,
    pub downs: i64,
    pub ups: i64,
    pub body_html: String,
    pub subreddit: String,
    pub name: String,
    pub score_hidden: bool,
    pub stickied: bool,
    pub created: i64,
    pub author_flair_text: Option<String>,
    pub created_utc: i64,
    pub distinguished: Option<String>,
    pub num_reports: Option<u64>,
    pub parent_id: String,
}

A deserializable structure representing a comment. This is created when the client returns JSON representing a comment and this is wrapped in a models::comment::Comment for ease-of-use.

Fields

subreddit_id

The Reddit ID for the subreddit where this was posted, including the leading t5_.

banned_by

Contains the name of the moderator who banned this, if the logged-in user is a moderator of this subreddit and this is banned.

removal_reason

The reason for the post removal, if you are a moderator and this post has been removed.

link_id

The ID (including kind) of the parent post (i.e. the link post or self post).

likes

This is Some(true) if the logged-in user has upvoted this submission, Some(false) if the user has downvoted this submission or None if the user has not voted.

replies

A listing of replies to this comment. Do not access this directly - use the functions of Comment instead.

saved

True if the logged-in user has saved this submission.

id

The ID of the post in base-36 form, as used in Reddit's links.

gilded

The amount of times that a user has been gilded (gifted Reddit Gold).

archived

This is true if Reddit has archived the submission (usually done after 6 months). Archived submissions cannot be voted or commented upon.

author

The name of the author of the submission (not including the leading /u/)

score

The overall points score of this post, as shown on the upvote counter. This is the same as upvotes - downvotes (however, this figure may be fuzzed by Reddit, and may not be exact)

approved_by

This contains the name of the user who approved this submission. This is None unless you are a mod of the subreddit and a user has approved this post.

body

The main post text.

edited

This is false if the submission is not edited and is the edit timestamp if it is edited. Access through the functions of Submission instead.

author_flair_css_class

The CSS class set for the author's flair (if available). If there is no flair, this is None.

downs

The number of downvotes (fuzzed; see score for further explanation)

ups

The number of upvotes (fuzzed; see score for further explanation)

body_html

Contains the HTML of the post body.

subreddit

The subreddit that this submission was posted in (not including /r/)

name

The full 'Thing ID', consisting of a 'kind' and a base-36 identifier. The valid kinds are: - t1_ - Comment - t2_ - Account - t3_ - Link - t4_ - Message - t5_ - Subreddit - t6_ - Award - t8_ - PromoCampaign

score_hidden

true if the score should not be displayed.

stickied

This is true if this submission is stickied (an 'annoucement' thread)

created

A timestamp of the time when the post was created, in the logged-in user's local time.

author_flair_text

The text of the author's flair, if present. Can be an empty string if the flair is present but contains no text.

created_utc

A timestamp of the time when the post was created, in UTC.

distinguished

Indicates whether the user has used a special flag for themselves, e.g. [M] or [A]. Possible values: - None - Normal user - Some("moderator") - [M] - Some("admin") - [A] - Some("special") - other special 'distinguishes' e.g. [Δ]

num_reports

parent_id