Module rawr::responses::listing

Structs

ListingData
The contents of a call to a 'listing' endpoint.

Submission
Represents all types of link posts and self posts on Reddit.

SubredditAboutData

Type Definitions

CommentResponse
The structure returned by the Reddit API for comments, consisting of the original post and a listing of the comments.

Listing
The 'listing' format returned by the Reddit API for post lists.

SubredditAbout
API response from /r/subreddit/about