Struct rawr::structures::comment::Comment

pub struct Comment<'a> {
    // some fields omitted
}

Structure representing a comment and its associated data (e.g. replies)

Methods

impl<'a> Comment<'a>

fn new(client: &RedditClient, data: _Comment) -> Comment

Internal method. Use Submission.replies() or Comment.replies() to get a listing, then select the desired comment instead.

fn parent(&self) -> &str

Gets the full ID of the parent submission/comment (kind + id e.g. 't1_4te6jf')

fn add_reply(&mut self, item: Comment<'a>)

Adds a reply to this comment's reply list. This is an internal method - to make the client reply to this post, use Comment.reply(MESSAGE).