Module rawr::structures::listing

Paginated list of submissions.

Structs

Listing
A paginated listing of posts that can be iterated through. Posts are fetched lazily until the listing is exhausted (similar to an infinite scroll of posts).

PostStream
An infinite stream of posts which updates as posts are received. Stores a list of seen posts so that each post is only seen once.