Struct rawr::structures::messages::MessageInterface

pub struct MessageInterface<'a> {
    // some fields omitted
}

A helper struct which allows access to the inbox, unread messages and other message queues.

Methods

impl<'a> MessageInterface<'a>

fn new(client: &RedditClient) -> MessageInterface

Internal method. Use RedditClient.messages() instead.

fn inbox(&self, opts: ListingOptions) -> Result<MessageListing<'a>, APIError>

Gets a list of all received messages that have not been deleted.

fn unread(&self, opts: ListingOptions) -> Result<MessageListing<'a>, APIError>

Gets all messages that have not been marked as read.

fn unread_stream(self) -> MessageStream<'a>

Gets a MessageStream of unread posts, marking each one as read after yielding it from the iterator. This can be useful to monitor /u/username mentions, replies to comments/posts and private messages.

Examples

use rawr::prelude::*;
let client = RedditClient::new("rawr", PasswordAuthenticator::new("a", "b", "c", "d"));
for message in client.messages().unread_stream() {
    println!("New message received.");
}