Module rawr::structures::submission

Structure representing a link post or self post (a submission).

Structs

FlairList
A list of flairs that can be assigned to a post. To access the complete list, use FlairList.flairs, which is a list of FlairChoice objects.

LazySubmission
A lazy object representing a submission. Used by the Client.get_by_id() method until the data is specified by the user (we don't know if they want the Submission or CommentList yet). The LazySubmission object is consumed when performing either of these actions.

Submission
Structure representing a link post or self post (a submission) on Reddit.