Struct rawr::structures::submission::FlairList

pub struct FlairList {
    pub flairs: Vec<FlairChoice>,
}

A list of flairs that can be assigned to a post. To access the complete list, use FlairList.flairs, which is a list of FlairChoice objects.

Fields

flairs

The list of flairs available.

Methods

impl FlairList

fn new(choices: Vec<FlairChoice>) -> FlairList

Creates a FlairList from a vector of FlairChoice objects.

fn find_text(self, text: &str) -> Option<String>

Finds the flair with the specified text, consuming the FlairList.

Examples

use rawr::client::RedditClient;
use rawr::auth::PasswordAuthenticator;
use rawr::options::ListingOptions;
use rawr::traits::Flairable;
let client = RedditClient::new("rawr", PasswordAuthenticator::new("a", "b", "c", "d"));
let sub = client.subreddit("learnprogramming");
let post = sub.hot(ListingOptions::default()).unwrap().next().unwrap();
// NOTE: this would 403 unless you are a moderator or the creator of the post.
let tutorial_flair = post.flair_options().unwrap().find_text("tutorial").unwrap();
post.flair(&tutorial_flair);