Struct rawr::structures::submission::LazySubmission

pub struct LazySubmission<'a> {
    // some fields omitted
}

A lazy object representing a submission. Used by the Client.get_by_id() method until the data is specified by the user (we don't know if they want the Submission or CommentList yet). The LazySubmission object is consumed when performing either of these actions.

Methods

impl<'a> LazySubmission<'a>

fn new(client: &'a RedditClient, id: &str) -> LazySubmission<'a>

Internal method. Use Client.get_by_id() instead.

fn get(self) -> Result<Submission<'a>, APIError>

Fetches the Submission with this ID, in order to access post title, body, link and creation time.

fn replies(self) -> Result<CommentList<'a>, APIError>

Fetches a CommentList with replies to this submission.