Struct rawr::structures::submission::Submission

pub struct Submission<'a> {
    // some fields omitted
}

Structure representing a link post or self post (a submission) on Reddit.

Methods

impl<'a> Submission<'a>

fn new(client: &RedditClient, data: Submission) -> Submission

Internal method. Get submissions from a listing instead (see Subreddit.hot() etc.)

fn reply_stream(self) -> CommentStream<'a>

Returns a CommentStream that fetches the latest comments in an infinite loop and returns it from the iterator. Comments will be ordered from oldest to newest, with up to 5 comments that exist being yielded at a time. This will poll the API every 5 seconds for updates.

Examples

use rawr::prelude::*;
let client = RedditClient::new("rawr", AnonymousAuthenticator::new());
let sub = client.subreddit("all");
let mut listing = sub.hot(ListingOptions::default()).expect("Could not fetch listing!");
let post = listing.nth(0).unwrap();
for comment in post.reply_stream() {
    println!("New comment received!");
}

fn title(&self) -> &str

The title of the post (as an &str). All link and self posts have a title, and any post flairs are not included in this.

fn is_self_post(&self) -> bool

This is true if the post is a self post, and false if it is a link post.

fn link_url(&self) -> Option<String>

Gets the URL linked to by this link post (or None, if this is a self post)

fn nsfw(&self) -> bool

Returns true if the post is marked NSFW (over 18).

fn mark_nsfw(&mut self) -> Result<(), APIError>

Sets the post as NSFW (over 18) if you have the correct privileges (owner of the post or moderator) and the subreddit allows NSFW posts.

fn unmark_nsfw(&mut self) -> Result<(), APIError>

Sets the post as not NSFW (over 18).