Struct rawr::structures::subreddit::Subreddit

pub struct Subreddit<'a> {
    pub name: String,
    // some fields omitted
}

The Subreddit struct represents a subreddit and allows access to post listings and data about the subreddit.

Fields

name

The name of the subreddit represented by this struct.

Methods

impl<'a> Subreddit<'a>

fn create_new(client: &'a RedditClient, name: &str) -> Subreddit<'a>

Creates a Subreddit from a client and the subreddit's name. Do not use this directly - use Client.subreddit(NAME) instead.

fn hot(&self, opts: ListingOptions) -> Result<Listing, APIError>

Gets a listing of the hot feed for this subreddit. The first page may include some sticky posts in addtion to the expected posts.

Examples

use rawr::client::RedditClient;
use rawr::options::ListingOptions;
use rawr::auth::AnonymousAuthenticator;
let client = RedditClient::new("rawr", AnonymousAuthenticator::new());
let sub = client.subreddit("askreddit");
let hot = sub.hot(ListingOptions::default());

fn new_stream(self) -> PostStream<'a>

Gets a PostStream of the new posts in the subreddit. The iterator will yield values forever, unless it is manually ended at some point. For tips, check the PostStream class.

Examples

use rawr::prelude::*;
let client = RedditClient::new("rawr", AnonymousAuthenticator::new());
let askreddit = client.subreddit("askreddit");
for post in askreddit.new_stream() {

}

fn new(&self, opts: ListingOptions) -> Result<Listing, APIError>

Gets a listing of the new feed for this subreddit.

Examples

use rawr::client::RedditClient;
use rawr::options::ListingOptions;
use rawr::traits::Content;
use rawr::auth::AnonymousAuthenticator;
let client = RedditClient::new("rawr", AnonymousAuthenticator::new());
let sub = client.subreddit("programming");
let mut new = sub.new(ListingOptions::default()).expect("Could not get new feed");
assert_eq!(new.next().unwrap().subreddit().name, "programming");

fn rising(&self, opts: ListingOptions) -> Result<Listing, APIError>

Gets a listing of the rising feed for this subreddit. Usually much shorter than the other listings; may be empty.

Examples

use rawr::client::RedditClient;
use rawr::options::ListingOptions;
use rawr::auth::AnonymousAuthenticator;
let client = RedditClient::new("rawr", AnonymousAuthenticator::new());
let sub = client.subreddit("thanksobama");
let rising = sub.rising(ListingOptions::default()).unwrap();
assert_eq!(rising.count(), 0);

fn top(&self, opts: ListingOptions, time: TimeFilter) -> Result<Listing, APIError>

Gets a listing of the top feed for this subreddit. Also requires a time filter ( rawr::options::TimeFilter) which is equivalent to the "links from: all time" dropdown on the website.

Examples

use rawr::client::RedditClient;
use rawr::options::{ListingOptions, TimeFilter};
use rawr::auth::AnonymousAuthenticator;
let client = RedditClient::new("rawr", AnonymousAuthenticator::new());
let sub = client.subreddit("thanksobama");
let mut top = sub.top(ListingOptions::default(), TimeFilter::AllTime)
    .expect("Request failed");
assert_eq!(top.next().unwrap().title(), "Thanks Me");

fn controversial(&self, opts: ListingOptions, time: TimeFilter) -> Result<Listing, APIError>

Gets a listing of the controversial feed for this subreddit. Also requires a time filter ( rawr::options::TimeFilter) which is equivalent to the "links from: all time" dropdown on the website.

fn submit_link(&self, post: LinkPost) -> Result<(), APIError>

Submits a link post to this subreddit using the specified parameters. If the link has already been posted, this will fail unless you specifically allow reposts.

Examples

Allowing a link to be reposted

use rawr::options::LinkPost;
let post = LinkPost::new("rawr!", "http://example.com").resubmit();

Submitting a post

use rawr::auth::PasswordAuthenticator;
use rawr::client::RedditClient;
use rawr::options::LinkPost;
let client = RedditClient::new("rawr", PasswordAuthenticator::new("a", "b", "c", "d"));
let sub = client.subreddit("rust");
let post = LinkPost::new("rawr!", "http://example.com");
sub.submit_link(post).expect("Posting failed!");

fn submit_text(&self, post: SelfPost) -> Result<(), APIError>

Submits a text post (self post) to this subreddit using the specified title and body.

Examples

Submitting a post

use rawr::auth::PasswordAuthenticator;
use rawr::client::RedditClient;
use rawr::options::SelfPost;
let client = RedditClient::new("rawr", PasswordAuthenticator::new("a", "b", "c", "d"));
let sub = client.subreddit("rust");
let post = SelfPost::new("I love rawr!", "You should download it *right now*!");
sub.submit_text(post).expect("Posting failed!");

fn about(&self) -> Result<SubredditAbout, APIError>

Fetches information about a subreddit such as subscribers, active users and sidebar information.

Examples

use rawr::client::RedditClient;
use rawr::auth::AnonymousAuthenticator;
let client = RedditClient::new("rawr", AnonymousAuthenticator::new());
let learn_programming = client.subreddit("learnprogramming").about()
    .expect("Could not fetch 'about' data");
assert_eq!(learn_programming.display_name(), "learnprogramming");