Struct rawr::structures::subreddit::SubredditAbout

pub struct SubredditAbout {
    // some fields omitted
}

Information about a subreddit such as subscribers, sidebar text and active users.

Methods

impl SubredditAbout

fn new(data: SubredditAboutData) -> SubredditAbout

Creates a new SubredditAbout instance. Use Subreddit.about() instead to get information about a subreddit.

fn subscribers(&self) -> u64

The number of subscribers to this subreddit.

fn active_users(&self) -> u64

The number of logged-in users who have viewed this subreddit in the last 15 minutes.

fn public(&self) -> bool

Returns true if the subreddit is visible to the public (i.e. not invitation only)

fn display_name(&self) -> &str

The display name of the subreddit, not including leading /r/