Struct rawr::structures::user::User

pub struct User<'a> {
    pub name: String,
    // some fields omitted
}

Interface to a Reddit user, which can be used to access their karma and moderator status.

Fields

name

The name of the user that this struct represents.

Methods

impl<'a> User<'a>

fn new(client: &'a RedditClient, name: &str) -> User<'a>

Internal method. Use RedditClient.user(NAME) instead.

fn about(self) -> Result<UserAbout, APIError>

Gets information about this user.

Example

use rawr::client::RedditClient;
use rawr::auth::AnonymousAuthenticator;
let client = RedditClient::new("rawr", AnonymousAuthenticator::new());
let user = client.user("Aurora0001").about().expect("User request failed");
assert_eq!(user.id(), "eqyvc");

fn flair_options(&self, subreddit: &str) -> Result<FlairList, APIError>

Gets a list of possible user flairs that can be added in this subreddit.

User flairs apply on a per-subreddit basis, and some may not permit user flairs at all. If you do not have the privileges to set the flair for this user, you will receive a 403 error.

fn flair(&self, subreddit: &str, template: &str) -> Result<(), APIError>

Sets the flair for this user in the specified subreddit, using the specified template string. You can get the template string from flair_options; either: - use the returned FlairList and call the method find_text which will return the template ID of the flair with the specified text. - iterate through the FlairList, and get the FlairChoice.flair_template_id value.

fn submissions(&self) -> Result<Listing, APIError>

Gets a list of submissions that the specified user has submitted. This endpoint is a listing and will continue yielding items until every item has been exhausted.

Examples

use rawr::prelude::*;
let client = RedditClient::new("rawr", AnonymousAuthenticator::new());
let user = client.user("Aurora0001");
let submissions = user.submissions().expect("Could not fetch!");
let mut i = 0;
for submission in submissions.take(5) {
    i += 1;
}
assert_eq!(i, 5);