Module rawr::traits

Traits for common API functionality.

Traits

Commentable
An object that can be commented upon and may have comments.

Content
An object that was created by an author and is in a subreddit (i.e. a submission or comment)

Created
An object that was created at some point (e.g. a subreddit, a submission or a comment)

Distinguishable
An object that can be distinguished (moderator/admin/special indicator).

Editable
An object that can be edited (anything that has a body).

Flairable
An object that has a flair and can be given a flair by the creator or a moderator.

Lockable
An object that can be locked so that no further comments can be added.

PageListing
A paginatable listing.

Reportable
An object that can be reported (submission, comment, private message). Also exposes moderation options to review reports.

Stickable
An object that can be stickied (made into an 'annoucement post'). In practice, this is only self/link posts.

Visible
An object that can be shown/hidden in listings.

Votable
An object that can be voted upon and has a score based on the upvotes - downvotes.